:building_construction: Refactoring scent code
