+++ app/src/main/java/com/mashup/lastgarden/data/repository/StoryRepository.kt
@@ -2,7 +2,7 @@ package com.mashup.lastgarden.data.repository
 
 import com.mashup.lastgarden.data.remote.StoryRemoteDataSource
 import com.mashup.lastgarden.data.vo.Story
-import com.mashup.lastgarden.network.services.LikeResponse
+import com.mashup.lastgarden.network.response.LikeResponse
 import javax.inject.Inject
 
 class StoryRepository @Inject constructor(private val remote: StoryRemoteDataSource) {
