+++ app/src/main/java/com/mashup/lastgarden/network/response/LikeResponse.kt
@@ -1,0 +1,5 @@
+package com.mashup.lastgarden.network.response
+
+data class LikeResponse(
+    val isLike: Boolean
+)
