+++ app/src/main/java/com/mashup/lastgarden/utils/formatNumber.kt
@@ -1,0 +1,8 @@
+package com.mashup.lastgarden.utils
+
+import java.text.DecimalFormat
+
+fun formatNumber(number: Long): String {
+    val formatter = DecimalFormat("###,###")
+    return formatter.format(number)
+}
